refactor(cli): split graph.ts into graph-builder.ts and graph-format.ts (sl-3fs7)
