+++ tooling/satsuma-cli/src/commands/graph-builder.ts
@@ -1,0 +1,618 @@
+/**
+ * graph-builder.ts — workspace graph assembly for `satsuma graph`
+ *
+ * Owns the construction of the full workspace semantic graph: node collection,
+ * schema-level edge derivation, and field-level edge resolution. The output
+ * is a plain data structure (WorkspaceGraph) consumed by the graph command's
+ * formatters and JSON serialiser.
+ *
+ * Does NOT own CLI registration, option parsing, or output formatting — those
+ * live in graph.ts and graph-format.ts respectively.
+ */
+
+import { resolve } from "node:path";
+import { canonicalKey } from "../index-builder.js";
+import { expandEntityFields, expandNestedSpreads } from "../spread-expand.js";
+import { extractAtRefs, classifyRef, resolveRef, resolveAllNLRefs } from "../nl-ref-extract.js";
+import type { WorkspaceIndex, FieldDecl } from "../types.js";
+import type { FullGraph } from "../graph-builder.js";
+
+// ── Public types ─────────────────────────────────────────────────────────────
+
+/** A schema-level directed edge (source → mapping, mapping → target, etc.). */
+export interface SchemaEdge {
+  from: string;
+  to: string;
+  /** Role describing how the edge connects: source, target, metric_source, or nl_ref. */
+  role: string;
+}
+
+/** A field-level directed edge representing one arrow in a mapping body. */
+export interface FieldEdge {
+  /** Fully-qualified source field path, or null for derived/generated arrows. */
+  from: string | null;
+  /** Fully-qualified target field path, or null for source-only arrows. */
+  to: string | null;
+  /** The mapping that owns this arrow. */
+  mapping: string;
+  /** Arrow classification: "none" | "structural" | "nl" | "mixed" | "nl-derived". */
+  classification: string;
+  /** Source file containing the arrow. */
+  file: string;
+  /** 1-based line number. */
+  line: number;
+  /** Structural transform steps (present for structural/mixed arrows). */
+  transforms?: string[];
+  /** Raw NL text from quoted pipe steps (present for nl/mixed arrows). */
+  nl_text?: string;
+  /** True when the arrow was generated by a `derived` block. */
+  derived?: boolean;
+}
+
+/** Options controlling graph assembly, derived from CLI flags. */
+export interface GraphBuildOpts {
+  /** Filter nodes and edges to this namespace only. */
+  namespace?: string | null;
+  /** Whether to include NL text in edges (false when --no-nl). */
+  includeNl: boolean;
+  /** Whether to omit field-level detail (--schema-only). */
+  schemaOnly: boolean;
+}
+
+/** The complete workspace graph data structure emitted by --json. */
+export interface WorkspaceGraph {
+  /** Schema version for the graph output format. */
+  version: number;
+  /** ISO-8601 timestamp of when the graph was generated. */
+  generated: string;
+  /** Absolute path to the workspace root. */
+  workspace: string;
+  /** Aggregate counts of entities and errors. */
+  stats: {
+    schemas: number;
+    mappings: number;
+    metrics: number;
+    fragments: number;
+    transforms: number;
+    arrows: number;
+    errors: number;
+  };
+  /** All graph nodes (schemas, mappings, metrics, transforms). */
+  nodes: Array<Record<string, unknown>>;
+  /** Field-level edges (or schema-level aggregated edges when --schema-only). */
+  edges: FieldEdge[];
+  /** Schema-level topology edges. */
+  schema_edges: SchemaEdge[];
+  /** Parser warnings surfaced from the workspace index. */
+  warnings: Array<{ text: string; file: string; line: number }>;
+  /** NL references that could not be resolved to a known schema or field. */
+  unresolved_nl: Array<{ scope: string; arrow: string; text: string; file: string; line: number }>;
+}
+
+// ── Graph construction ───────────────────────────────────────────────────────
+
+/**
+ * Return an empty graph for workspaces with no .stm files.
+ * Ensures --json always emits a structurally valid object.
+ */
+export function emptyGraph(root: string): WorkspaceGraph {
+  return {
+    version: 1,
+    generated: new Date().toISOString(),
+    workspace: resolve(root),
+    stats: { schemas: 0, mappings: 0, metrics: 0, fragments: 0, transforms: 0, arrows: 0, errors: 0 },
+    nodes: [],
+    edges: [],
+    schema_edges: [],
+    warnings: [],
+    unresolved_nl: [],
+  };
+}
+
+/**
+ * Build the full workspace graph output structure from a parsed workspace
+ * index and its schema-level directed graph.
+ *
+ * Collects nodes (schemas, mappings, metrics, transforms), builds schema-level
+ * and field-level edges, applies namespace filtering, and computes stats.
+ */
+export function buildWorkspaceGraph(index: WorkspaceIndex, schemaGraph: FullGraph, root: string, opts: GraphBuildOpts): WorkspaceGraph {
+  const nsFilter = opts.namespace ?? null;
+  const includeNl = opts.includeNl;
+  const schemaOnly = opts.schemaOnly;
+
+  // ── Collect nodes ──────────────────────────────────────────────────────────
+  const nodes: Array<Record<string, unknown>> = [];
+  const includedNodeIds = new Set<string>();
+
+  for (const [id, schema] of index.schemas) {
+    if (nsFilter && schema.namespace !== nsFilter) continue;
+    includedNodeIds.add(id);
+    const node: Record<string, unknown> = {
+      id,
+      kind: "schema",
+      namespace: schema.namespace ?? null,
+      file: schema.file,
+      line: schema.row + 1,
+      note: schema.note ?? null,
+    };
+    if (!schemaOnly) {
+      const fieldsCopy: FieldDecl[] = JSON.parse(JSON.stringify(schema.fields)) as FieldDecl[];
+      expandNestedSpreads(fieldsCopy, schema.namespace ?? null, index);
+      const spreadFields = expandEntityFields(schema, schema.namespace ?? null, index);
+      node.fields = ([...fieldsCopy, ...spreadFields] as FieldDecl[]).map((f) => ({
+        name: f.name,
+        type: f.isList && f.type ? `list_of ${f.type}` : (f.type ?? null),
+      }));
+    }
+    nodes.push(node);
+  }
+
+  for (const [id, mapping] of index.mappings) {
+    if (nsFilter && mapping.namespace !== nsFilter) continue;
+    includedNodeIds.add(id);
+    nodes.push({
+      id,
+      kind: "mapping",
+      namespace: mapping.namespace ?? null,
+      file: mapping.file,
+      line: mapping.row + 1,
+      sources: mapping.sources,
+      targets: mapping.targets,
+    });
+  }
+
+  for (const [id, metric] of index.metrics) {
+    if (nsFilter && metric.namespace !== nsFilter) continue;
+    includedNodeIds.add(id);
+    const metricNode: Record<string, unknown> = {
+      id,
+      kind: "metric",
+      namespace: metric.namespace ?? null,
+      file: metric.file,
+      line: metric.row + 1,
+      sources: metric.sources,
+      grain: metric.grain ?? null,
+      slices: metric.slices ?? [],
+    };
+    if (!schemaOnly) {
+      metricNode.fields = metric.fields.map((f) => ({
+        name: f.name,
+        type: f.isList && f.type ? `list_of ${f.type}` : (f.type ?? null),
+      }));
+    }
+    nodes.push(metricNode);
+  }
+
+  // Fragments are macro definitions — they do not appear as graph nodes.
+  // Their fields are expanded into the consuming schema's field list via
+  // expandEntityFields above. stats.fragments still reports the count.
+
+  for (const [id, transform] of index.transforms) {
+    if (nsFilter && transform.namespace !== nsFilter) continue;
+    includedNodeIds.add(id);
+    nodes.push({
+      id,
+      kind: "transform",
+      namespace: transform.namespace ?? null,
+      file: transform.file,
+      line: transform.row + 1,
+    });
+  }
+
+  // ── Build edges ────────────────────────────────────────────────────────────
+  const schemaEdges = buildSchemaEdges(index, schemaGraph, includedNodeIds, nsFilter);
+
+  // Always build field edges (needed for --schema-only aggregation too)
+  const result = buildFieldEdges(index, includedNodeIds, nsFilter, includeNl);
+
+  const unresolvedNl: Array<{ scope: string; arrow: string; text: string; file: string; line: number }> = [];
+
+  const fieldEdges: FieldEdge[] = schemaOnly
+    ? aggregateFieldEdgesToSchemaLevel(result.edges, index, nsFilter)
+    : result.edges;
+  unresolvedNl.push(...result.unresolvedNl);
+
+  // Count arrows (raw field arrows, not aggregated)
+  const arrowCount = result.edges.length;
+
+  // ── Assemble output ────────────────────────────────────────────────────────
+  return {
+    version: 1,
+    generated: new Date().toISOString(),
+    workspace: resolve(root),
+    stats: {
+      schemas: [...index.schemas.values()].filter((s) => !nsFilter || s.namespace === nsFilter).length,
+      mappings: [...index.mappings.values()].filter((m) => !nsFilter || m.namespace === nsFilter).length,
+      metrics: [...index.metrics.values()].filter((m) => !nsFilter || m.namespace === nsFilter).length,
+      fragments: [...index.fragments.values()].filter((f) => !nsFilter || f.namespace === nsFilter).length,
+      transforms: [...index.transforms.values()].filter((t) => !nsFilter || t.namespace === nsFilter).length,
+      arrows: arrowCount,
+      errors: index.totalErrors,
+    },
+    nodes,
+    edges: fieldEdges,
+    schema_edges: schemaEdges,
+    warnings: index.warnings.map((w) => ({ text: w.text, file: w.file, line: w.row + 1 })),
+    unresolved_nl: includeNl ? unresolvedNl : [],
+  };
+}
+
+// ── Schema-level edges ───────────────────────────────────────────────────────
+
+/**
+ * Build schema-level edges from the directed graph.
+ * Each edge has: from, to, role (source/target/metric_source/nl_ref).
+ */
+function buildSchemaEdges(index: WorkspaceIndex, _schemaGraph: FullGraph, includedNodeIds: Set<string>, nsFilter: string | null): SchemaEdge[] {
+  const edges: SchemaEdge[] = [];
+
+  // ── Mapping source/target edges ────────────────────────────────────────────
+  for (const [mappingName, mapping] of index.mappings) {
+    // When namespace-filtering, include edges if either the mapping is in
+    // the namespace OR any of its sources/targets are in the namespace
+    if (nsFilter && mapping.namespace !== nsFilter) {
+      const touchesNs = mapping.sources.some((s) => includedNodeIds.has(s)) ||
+        mapping.targets.some((t) => includedNodeIds.has(t));
+      if (!touchesNs) continue;
+    }
+
+    for (const src of mapping.sources) {
+      if (!nsFilter || includedNodeIds.has(src) || includedNodeIds.has(mappingName)) {
+        edges.push({ from: src, to: mappingName, role: "source" });
+      }
+    }
+    for (const tgt of mapping.targets) {
+      if (!nsFilter || includedNodeIds.has(tgt) || includedNodeIds.has(mappingName)) {
+        edges.push({ from: mappingName, to: tgt, role: "target" });
+      }
+    }
+  }
+
+  // ── Metric source edges ────────────────────────────────────────────────────
+  for (const [metricName, srcSchemas] of index.referenceGraph.metricsReferences) {
+    const metric = index.metrics.get(metricName);
+    if (nsFilter && metric?.namespace !== nsFilter) continue;
+
+    for (const src of srcSchemas) {
+      if (!nsFilter || includedNodeIds.has(src) || includedNodeIds.has(metricName)) {
+        edges.push({ from: src, to: metricName, role: "metric_source" });
+      }
+    }
+  }
+
+  // ── NL @ref edges ──────────────────────────────────────────────────────────
+  // Promote resolved NL schema references to first-class edges.
+  // Now that hidden-source-in-nl is an error (P5.1), these references are
+  // guaranteed to be declared in the mapping's source list, so they represent
+  // intentional data lineage, not phantom paths.
+  if (index.nlRefData) {
+    const seen = new Set<string>();
+    for (const item of index.nlRefData) {
+      const mappingKey = item.namespace
+        ? `${item.namespace}::${item.mapping}`
+        : item.mapping;
+      const mapping = index.mappings.get(mappingKey);
+      if (!mapping) continue;
+
+      const backtickRefs = extractNlSchemaRefs(item.text, {
+        sources: mapping.sources ?? [],
+        targets: mapping.targets ?? [],
+        namespace: item.namespace,
+      }, index);
+
+      for (const schemaRef of backtickRefs) {
+        const key = `${schemaRef}|${mappingKey}|nl_ref`;
+        if (seen.has(key)) continue;
+        seen.add(key);
+        if (!nsFilter || includedNodeIds.has(schemaRef) || includedNodeIds.has(mappingKey)) {
+          edges.push({ from: schemaRef, to: mappingKey, role: "nl_ref" });
+        }
+      }
+    }
+  }
+
+  return edges;
+}
+
+// ── NL schema ref extraction ─────────────────────────────────────────────────
+
+/**
+ * Extract index-key-format schema names referenced via @refs in NL text.
+ * Filters out schemas already in the mapping's declared sources/targets (those
+ * are already covered by the standard source/target edges).
+ */
+function extractNlSchemaRefs(
+  text: string,
+  mappingContext: { sources: string[]; targets: string[]; namespace: string | null },
+  index: WorkspaceIndex,
+): string[] {
+  const refs = extractAtRefs(text);
+  const schemas: string[] = [];
+  const allDeclared = new Set([...mappingContext.sources, ...mappingContext.targets]);
+
+  for (const { ref } of refs) {
+    const classification = classifyRef(ref);
+    const resolution = resolveRef(ref, mappingContext, index);
+    if (!resolution.resolved) continue;
+
+    let canonicalSchema: string | null = null;
+    if (classification === "namespace-qualified-schema" || classification === "bare") {
+      if (resolution.resolvedTo?.kind === "schema") {
+        canonicalSchema = resolution.resolvedTo.name;
+      }
+    } else if (classification === "dotted-field" || classification === "namespace-qualified-field") {
+      if (resolution.resolvedTo?.kind === "field") {
+        const fieldName = resolution.resolvedTo.name;
+        const lastDot = fieldName.lastIndexOf(".");
+        if (lastDot > 0) {
+          canonicalSchema = fieldName.slice(0, lastDot);
+        }
+      }
+    }
+
+    if (!canonicalSchema) continue;
+
+    // Convert canonical form (::name) to index key form (name)
+    const indexKey = canonicalSchema.startsWith("::")
+      ? canonicalSchema.slice(2)
+      : canonicalSchema;
+
+    // Skip schemas already declared as source or target
+    if (allDeclared.has(indexKey) || allDeclared.has(canonicalSchema)) continue;
+
+    schemas.push(indexKey);
+  }
+
+  return schemas;
+}
+
+// ── Field-level edges ────────────────────────────────────────────────────────
+
+/**
+ * Qualify a field path with its schema name, handling edge cases:
+ *  - Leading dot (nested field, e.g. ".PHONE_TYPE") -> "schema.PHONE_TYPE"
+ *  - Already schema-qualified (multi-source, e.g. "crm.customer_id") -> unchanged
+ *    when the prefix matches a known schema
+ *  - Simple field name -> "schema.field"
+ */
+function qualifyField(field: string, schemas: string[]): string {
+  if (schemas.length === 0) return field;
+
+  // Nested field: strip leading dot, prepend first schema
+  if (field.startsWith(".")) {
+    return `${schemas[0]}.${field.slice(1)}`;
+  }
+
+  // Multi-source: field may already be schema-qualified (e.g. "crm.customer_id")
+  const dotIdx = field.indexOf(".");
+  if (dotIdx > 0) {
+    const prefix = field.slice(0, dotIdx);
+    if (schemas.includes(prefix)) {
+      // Already qualified — don't double-prefix
+      return field;
+    }
+    // Check if prefix matches the bare name of a namespace-qualified schema
+    for (const s of schemas) {
+      const nsIdx = s.indexOf("::");
+      const bare = nsIdx !== -1 ? s.slice(nsIdx + 2) : s;
+      if (bare === prefix) return field;
+    }
+  }
+
+  return `${schemas[0]}.${field}`;
+}
+
+/**
+ * Build field-level edges from arrow records in the workspace index.
+ *
+ * Iterates all arrow records, qualifies source/target field paths with their
+ * parent schema names, attaches transform and NL metadata, and resolves
+ * nl-derived edges from @ref mentions.
+ */
+function buildFieldEdges(index: WorkspaceIndex, includedNodeIds: Set<string>, nsFilter: string | null, includeNl: boolean): { edges: FieldEdge[]; unresolvedNl: Array<{ scope: string; arrow: string; text: string; file: string; line: number }> } {
+  const edges: FieldEdge[] = [];
+  const unresolvedNl: Array<{ scope: string; arrow: string; text: string; file: string; line: number }> = [];
+
+  // Iterate all arrow records via the fieldArrows index.
+  // fieldArrows duplicates records under multiple keys, so deduplicate here.
+  const seen = new Set<string>();
+
+  for (const [_key, records] of index.fieldArrows) {
+    for (const record of records) {
+      // Deduplicate — same arrow record may appear under multiple keys
+      const dedupKey = `${record.file}:${record.line}:${record.target}`;
+      if (seen.has(dedupKey)) continue;
+      seen.add(dedupKey);
+
+      const mappingKey = record.namespace
+        ? `${record.namespace}::${record.mapping}`
+        : (record.mapping ?? "");
+
+      if (nsFilter) {
+        const mapping = index.mappings.get(mappingKey);
+        if (mapping?.namespace !== nsFilter) {
+          // Still include if the mapping touches schemas in the namespace
+          const touchesNs = (mapping?.sources ?? []).some((s) => includedNodeIds.has(s)) ||
+            (mapping?.targets ?? []).some((t) => includedNodeIds.has(t));
+          if (!touchesNs) continue;
+        }
+      }
+
+      // Resolve source and target schema names
+      const mapping = index.mappings.get(mappingKey);
+      const sourceSchemas = mapping?.sources ?? [];
+      const targetSchemas = mapping?.targets ?? [];
+
+      const fromFields = record.sources.length > 0
+        ? record.sources.map((s) => canonicalKey(qualifyField(s, sourceSchemas)))
+        : [null];
+      const toField = record.target
+        ? canonicalKey(qualifyField(record.target, targetSchemas))
+        : null;
+
+      for (const fromField of fromFields) {
+      const edge: FieldEdge = {
+        from: fromField,
+        to: toField,
+        mapping: mappingKey,
+        classification: record.classification,
+        file: record.file,
+        line: record.line + 1,
+      };
+
+      if (record.classification === "structural" || record.classification === "mixed") {
+        edge.transforms = record.steps
+          .filter((s) => s.type !== "pipe_text" || !s.text.startsWith('"'))
+          .map((s) => s.text);
+      }
+
+      if ((record.classification === "nl" || record.classification === "mixed") && includeNl) {
+        const nlSteps = record.steps.filter(
+          (s) => s.type === "pipe_text" && (s.text.startsWith('"') || s.text.startsWith('"""')),
+        );
+        if (nlSteps.length > 0) {
+          edge.nl_text = nlSteps.map((s) => s.text).join(" ");
+        }
+      }
+
+      if (record.derived) {
+        edge.derived = true;
+      }
+
+      edges.push(edge);
+      } // end for fromFields
+
+      // Track unresolved NL for the output section
+      if (record.classification === "nl" || record.classification === "mixed") {
+        const nlSteps = record.steps.filter(
+          (s) => s.type === "pipe_text" && (s.text.startsWith('"') || s.text.startsWith('"""')),
+        );
+        if (nlSteps.length > 0) {
+          unresolvedNl.push({
+            scope: `mapping ${mappingKey}`,
+            arrow: `-> ${record.target ?? "?"}`,
+            text: nlSteps.map((s) => s.text).join(" "),
+            file: record.file,
+            line: record.line + 1,
+          });
+        }
+      }
+    }
+  }
+
+  // ── NL-derived field edges ─────────────────────────────────────────────────
+  // Add nl-derived field edges from NL @ref mentions.
+  // These represent implicit field lineage: `@schema.field` in a mapping
+  // transform body means that field is an implicit source for the target field.
+  const nlRefs = resolveAllNLRefs(index);
+  const nlSeen = new Set<string>();
+  for (const nlRef of nlRefs) {
+    if (!nlRef.resolved || !nlRef.resolvedTo || nlRef.resolvedTo.kind !== "field") continue;
+    if (!nlRef.targetField) continue; // skip notes without a target arrow
+
+    // nlRef.mapping is already the fully-qualified key (e.g. "ns::name")
+    // built by resolveAllNLRefs — no additional namespace prefixing needed.
+    const mappingKey = nlRef.mapping;
+    const mapping = index.mappings.get(mappingKey);
+    if (!mapping) continue;
+
+    if (nsFilter) {
+      if (mapping.namespace !== nsFilter) {
+        const touchesNs = mapping.sources.some((s) => includedNodeIds.has(s)) ||
+          mapping.targets.some((t) => includedNodeIds.has(t));
+        if (!touchesNs) continue;
+      }
+    }
+
+    const sourceField = nlRef.resolvedTo.name; // already canonical, e.g. "::s1.a"
+    const rawTarget = qualifyField(nlRef.targetField, mapping.targets);
+    const targetField = canonicalKey(rawTarget);
+
+    // Skip if this is a self-reference (field references itself in its own transform)
+    if (sourceField === targetField) continue;
+
+    // Skip if the same source->target->mapping nl-derived edge was already emitted
+    const dedupKey = `${sourceField}|${targetField}|${mappingKey}`;
+    if (nlSeen.has(dedupKey)) continue;
+    nlSeen.add(dedupKey);
+
+    // Skip if there's already a declared (non-nl-derived) arrow with the same
+    // source->target in the same mapping (e.g. `c -> d { "@s1.c is processed" }`
+    // — c is already the declared source, no need for a duplicate nl-derived edge).
+    const alreadyCovered = edges.some(
+      (e) => e.from === sourceField && e.to === targetField &&
+             e.mapping === mappingKey && e.classification !== "nl-derived",
+    );
+    if (alreadyCovered) continue;
+
+    edges.push({
+      from: sourceField,
+      to: targetField,
+      mapping: mappingKey,
+      classification: "nl-derived",
+      file: nlRef.file,
+      line: nlRef.line + 1,
+    });
+  }
+
+  return { edges, unresolvedNl };
+}
+
+// ── Schema-only aggregation ──────────────────────────────────────────────────
+
+/**
+ * Aggregate field-level edges into schema-level edges by extracting the
+ * schema prefix from dotted field paths and deduplicating. For mappings
+ * with no field edges (e.g. derived-only), adds edges from the declared
+ * source/target lists.
+ */
+function aggregateFieldEdgesToSchemaLevel(fieldEdges: FieldEdge[], index: WorkspaceIndex, nsFilter: string | null): FieldEdge[] {
+  const aggregated: FieldEdge[] = [];
+  const seen = new Set<string>();
+
+  for (const edge of fieldEdges) {
+    const fromSchema = edge.from ? edge.from.split(".")[0] : null;
+    const toSchema = edge.to ? edge.to.split(".")[0] : null;
+    if (fromSchema && toSchema) {
+      const key = `${fromSchema}->${toSchema}:${edge.mapping}`;
+      if (!seen.has(key)) {
+        seen.add(key);
+        aggregated.push({
+          from: fromSchema,
+          to: toSchema,
+          mapping: edge.mapping,
+          classification: edge.classification,
+          file: edge.file,
+          line: edge.line,
+        });
+      }
+    }
+  }
+
+  // For mappings with no field edges (e.g. derived-only), add schema-level
+  // edges from the declared source/target lists.
+  const mappingsWithEdges = new Set(aggregated.map((e) => e.mapping));
+  for (const [id, mapping] of index.mappings) {
+    if (nsFilter && mapping.namespace !== nsFilter) continue;
+    if (id && mappingsWithEdges.has(id)) continue;
+    for (const src of mapping.sources) {
+      for (const tgt of mapping.targets) {
+        const key = `${src}->${tgt}:${id}`;
+        if (!seen.has(key)) {
+          seen.add(key);
+          aggregated.push({
+            from: src,
+            to: tgt,
+            mapping: id,
+            classification: "none",
+            file: mapping.file,
+            line: mapping.row + 1,
+          });
+        }
+      }
+    }
+  }
+
+  return aggregated;
+}
